Move isValidThemeFilterTerm
