+++ client/my-sites/theme/theme-features-card.js
@@ -13,7 +13,7 @@ import { localize } from 'i18n-calypso';
 import { Card } from '@automattic/components';
 import QueryThemeFilters from 'components/data/query-theme-filters';
 import SectionHeader from 'components/section-header';
-import isValidThemeFilterTerm from 'state/selectors/is-valid-theme-filter-term';
+import { isValidThemeFilterTerm } from 'state/themes/selectors';
 
 const ThemeFeaturesCard = ( { isWpcomTheme, siteSlug, features, translate } ) => {
 	if ( isEmpty( features ) ) {
