+++ client/state/selectors/test/is-valid-theme-filter-term.js
@@ -6,7 +6,7 @@ import { expect } from 'chai';
 /**
  * Internal dependencies
  */
-import isValidThemeFilterTerm from 'state/selectors/is-valid-theme-filter-term';
+import { isValidThemeFilterTerm } from 'state/themes/selectors';
 import { state } from './fixtures/theme-filters';
 
 describe( 'isValidThemeFilterTerm()', () => {
